Non-specific API improvements (#15)
* handoff message cloning directly to ecCodes

* add test iter beyond none

* remove need for clone to find nearest

* (tests not working) remove need for clone for KeysIter

* fix tests for new KeysIterator

* add initial null pointer check

* refactor intermediate bindings

* add libc null check

* refactor out codes_nearest

* refactor out keys_iterator

* refactor out keyed_message

* codes_index module as a file

* refactor Key definition

* rustfmt

* remove unwraps in tests

* ignore failing tests in main branch
