+++ .github/workflows/rust-dev.yml
@@ -29,7 +29,7 @@ jobs:
         cargo clippy --features "experimental_index"
     - name: Build release
       run: |
-        cargo build --release --features "experimental_index"
+        cargo test --features "experimental_index" -- --include-ignored
 
   build-macos:
 
@@ -51,4 +51,4 @@ jobs:
         cargo clippy --features "experimental_index"
     - name: Build release
       run: |
-        cargo build --release --features "experimental_index"
+        cargo test --features "experimental_index" -- --include-ignored
