+++ src/codes_index.rs
@@ -4,7 +4,7 @@
 use crate::{
     codes_handle::SpecialDrop,
     errors::CodesError,
-    intermediate_bindings::codes_index::{
+    intermediate_bindings::{
         codes_index_add_file, codes_index_new, codes_index_read, codes_index_select_double,
         codes_index_select_long, codes_index_select_string,
     },
@@ -107,54 +107,56 @@ impl Drop for CodesIndex {
 
 #[cfg(test)]
 mod tests {
+    use anyhow::Result;
+
     use crate::codes_index::{CodesIndex, Select};
     use std::path::Path;
     #[test]
-    fn index_constructors() {
+    fn index_constructors() -> Result<()> {
         {
             let keys = vec!["shortName", "typeOfLevel", "level", "stepType"];
-            let index = CodesIndex::new_from_keys(&keys).unwrap();
+            let index = CodesIndex::new_from_keys(&keys)?;
             assert!(!index.pointer.is_null());
         }
         {
             let file_path = Path::new("./data/iceland-surface.idx");
-            let index = CodesIndex::read_from_file(file_path).unwrap();
+            let index = CodesIndex::read_from_file(file_path)?;
             assert!(!index.pointer.is_null());
         }
+
+        Ok(())
     }
 
     #[test]
-    fn index_destructor() {
+    fn index_destructor() -> Result<()> {
         let keys = vec!["shortName", "typeOfLevel", "level", "stepType"];
-        let index = CodesIndex::new_from_keys(&keys).unwrap();
+        let index = CodesIndex::new_from_keys(&keys)?;
 
-        drop(index)
+        drop(index);
+        Ok(())
     }
 
     #[test]
-    fn add_file() {
+    fn add_file() -> Result<()> {
         let keys = vec!["shortName", "typeOfLevel", "level", "stepType"];
-        let index = CodesIndex::new_from_keys(&keys).unwrap();
+        let index = CodesIndex::new_from_keys(&keys)?;
         let grib_path = Path::new("./data/iceland.grib");
-        let index = index.add_grib_file(grib_path).unwrap();
+        let index = index.add_grib_file(grib_path)?;
 
         assert!(!index.pointer.is_null());
+        Ok(())
     }
 
     #[test]
-    fn index_selection() {
+    fn index_selection() -> Result<()> {
         let file_path = Path::new("./data/iceland-surface.idx");
-        let index = CodesIndex::read_from_file(file_path)
-            .unwrap()
-            .select("shortName", "2t")
-            .unwrap()
-            .select("typeOfLevel", "surface")
-            .unwrap()
-            .select("level", 0)
-            .unwrap()
-            .select("stepType", "instant")
-            .unwrap();
+        let index = CodesIndex::read_from_file(file_path)?
+            .select("shortName", "2t")?
+            .select("typeOfLevel", "surface")?
+            .select("level", 0)?
+            .select("stepType", "instant")?;
 
         assert!(!index.pointer.is_null());
+        Ok(())
     }
 }
