+++ src/errors.rs
@@ -47,6 +47,24 @@ pub enum CodesError {
     ///Similar to [`CodesInternal::CodesNotFound`] and [`CodesInternal::CodesMissingKey`].
     #[error("The key is missing in present message")]
     MissingKey,
+
+    /// Returned when codes_handle_clone returns null pointer
+    /// indicating issues with cloning the message.
+    #[error("Cannot clone the message")]
+    CloneFailed,
+
+    /// Returned when codes_keys_iterator_new returns null pointer
+    #[error("Cannot create or manipulate keys iterator")]
+    KeysIteratorFailed,
+
+    /// This error can be returned by almost any function in the crate.
+    /// It is returned when null pointer was passed to ecCodes function
+    /// which cannot handle null pointers. This error may indicate both
+    /// bug in the implementation or incorrect usage of the crate.
+    /// This error could be a panic, but as the crate is not comprehensively tested
+    /// it cannot be guaranteed that the null pointer is not caused by the user's mistake.
+    #[error("Null pointer encountered where it should not be")]
+    NullPtr,
 }
 
 #[derive(Copy, Eq, PartialEq, Clone, Ord, PartialOrd, Hash, Error, Debug, FromPrimitive)]
