+++ src/intermediate_bindings/codes_index.rs
@@ -1,19 +1,14 @@
 #![allow(non_camel_case_types)]
 #![allow(clippy::module_name_repetitions)]
 
-use eccodes_sys::{codes_index, CODES_LOCK};
+use eccodes_sys::{codes_context, codes_index, CODES_LOCK};
+use num_traits::FromPrimitive;
 use std::{ffi::CString, ptr};
 
-#[cfg(target_os = "macos")]
-type _SYS_IO_FILE = eccodes_sys::__sFILE;
-
-#[cfg(not(target_os = "macos"))]
-type _SYS_IO_FILE = eccodes_sys::_IO_FILE;
-
-use eccodes_sys::{codes_context, codes_handle};
-use num_traits::FromPrimitive;
-
-use crate::errors::{CodesError, CodesInternal};
+use crate::{
+    errors::{CodesError, CodesInternal},
+    pointer_guard,
+};
 
 // all index functions are safeguarded by a lock
 // because there are random errors appearing when using the index functions concurrently
@@ -61,6 +56,8 @@ pub unsafe fn codes_index_add_file(
     index: *mut codes_index,
     filename: &str,
 ) -> Result<(), CodesError> {
+    pointer_guard::non_null!(index);
+
     let filename = CString::new(filename).unwrap();
 
     let _g = CODES_LOCK.lock().unwrap();
@@ -78,6 +75,8 @@ pub unsafe fn codes_index_select_long(
     key: &str,
     value: i64,
 ) -> Result<(), CodesError> {
+    pointer_guard::non_null!(index);
+
     let key = CString::new(key).unwrap();
 
     let _g = CODES_LOCK.lock().unwrap();
@@ -95,6 +94,8 @@ pub unsafe fn codes_index_select_double(
     key: &str,
     value: f64,
 ) -> Result<(), CodesError> {
+    pointer_guard::non_null!(index);
+
     let key = CString::new(key).unwrap();
 
     let _g = CODES_LOCK.lock().unwrap();
@@ -112,6 +113,8 @@ pub unsafe fn codes_index_select_string(
     key: &str,
     value: &str,
 ) -> Result<(), CodesError> {
+    pointer_guard::non_null!(index);
+
     let key = CString::new(key).unwrap();
     let value = CString::new(value).unwrap();
 
@@ -123,25 +126,4 @@ pub unsafe fn codes_index_select_string(
         return Err(err.into());
     }
     Ok(())
-}
-
-pub unsafe fn codes_handle_new_from_index(
-    index: *mut codes_index,
-) -> Result<*mut codes_handle, CodesError> {
-    let mut error_code: i32 = 0;
-
-    let _g = CODES_LOCK.lock().unwrap();
-    let codes_handle = eccodes_sys::codes_handle_new_from_index(index, &mut error_code);
-
-    // special case! codes_handle_new_from_index returns -43 when there are no messages left in the index
-    // this is also indicated by a null pointer, which is handled upstream
-    if error_code == -43 {
-        return Ok(codes_handle);
-    }
-
-    if error_code != 0 {
-        let err: CodesInternal = FromPrimitive::from_i32(error_code).unwrap();
-        return Err(err.into());
-    }
-    Ok(codes_handle)
 }
