+++ tests/handle.rs
@@ -1,40 +1,41 @@
 use std::{path::Path, thread};
 
-use eccodes::{CodesHandle, KeyType, ProductKind, FallibleStreamingIterator};
+use anyhow::{Context, Result};
+use eccodes::{CodesHandle, FallibleStreamingIterator, KeyType, ProductKind};
 
 #[test]
-fn thread_safety() {
-    thread::spawn(|| loop {
-        let file_path = Path::new("./data/iceland.grib");
+fn thread_safety() -> Result<()> {
+    thread::spawn(|| -> anyhow::Result<()> {
+        loop {
+            let file_path = Path::new("./data/iceland.grib");
 
-        let mut handle = CodesHandle::new_from_file(file_path, ProductKind::GRIB).unwrap();
-        let current_message = handle.next().unwrap().unwrap();
+            let mut handle = CodesHandle::new_from_file(file_path, ProductKind::GRIB)?;
+            let current_message = handle.next()?.context("Message not some")?;
 
-        for _ in 0..100 {
-            let _ = current_message.read_key("name").unwrap();
+            for _ in 0..100 {
+                let _ = current_message.read_key("name")?;
 
-            let str_key = current_message.read_key("name").unwrap();
+                let str_key = current_message.read_key("name")?;
 
-            match str_key.value {
-                KeyType::Str(_) => {}
-                _ => panic!("Incorrect variant of string key"),
+                match str_key.value {
+                    KeyType::Str(_) => {}
+                    _ => panic!("Incorrect variant of string key"),
+                }
+
+                assert_eq!(str_key.name, "name");
             }
 
-            assert_eq!(str_key.name, "name");
+            drop(handle);
         }
-
-        drop(handle);
     });
 
     for _ in 0..1000 {
         let file_path = Path::new("./data/iceland.grib");
 
-        let mut handle = CodesHandle::new_from_file(file_path, ProductKind::GRIB).unwrap();
-        let current_message = handle.next().unwrap().unwrap();
+        let mut handle = CodesHandle::new_from_file(file_path, ProductKind::GRIB)?;
+        let current_message = handle.next()?.context("Message not some")?;
 
-        let long_key = current_message
-            .read_key("numberOfPointsAlongAParallel")
-            .unwrap();
+        let long_key = current_message.read_key("numberOfPointsAlongAParallel")?;
 
         match long_key.value {
             KeyType::Int(_) => {}
@@ -45,4 +46,6 @@ fn thread_safety() {
 
         drop(handle);
     }
+
+    Ok(())
 }
